# src/main/java/com/previred/taskmanagementapi/model/request/LoginRequest.java
package com.previred.taskmanagementapi.model.request;


import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank(message = "Username is required")
        String username,

        @NotBlank(message = "Password is required")
        String password
) {}